Added some performance benchmark

+++ backend/src/models/response/suggestions.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Default)]
 pub struct SuggestionResponse {
-    pub time_taken: u128,
+    pub time_taken: f64,
     pub total_results: usize,
     pub suggestions: Vec<String>
 }
